feat(errors): move error response to internal package
